moved from book details package as it is more suitable place
